tests/wpt/web-platform-tests/workers/Worker_ErrorEvent_lineno.htm

<!DOCTYPE html>
<title> AbstractWorker ErrorEvent.lineno </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onerror = t.step_func_done(function(e) {
    assert_equals(e.lineno, 3);
  });
  worker.postMessage("Error Message");
});
</script>
